==> Project/02 Missing Data Imputation.ipynb <==
# <center> **Kaggle’s Spaceship Titanic Competition**
# <center> **Imputing Missing Values**

# **Libraries**

import pandas as pd
import numpy as np
import seaborn as sns
import matplotlib as mpl
import matplotlib.pyplot as plt

import functions
import importlib
importlib.reload(functions)

from sklearn.preprocessing import LabelEncoder

import warnings

# **Load Data**

data = pd.read_csv(
    r"C:\Users\Dell\Documents\AI\Titanic\Data\data.csv",
    index_col=False
)

random_state = 101
target = 'Transported'

# data = data.drop([target], axis=1)

# **Display Parameters**

size = 20

params = {
    "font.family": "Times New Roman",
    "font.size": size,
    "axes.labelsize": size,
    "xtick.labelsize": size * 0.75,
    "ytick.labelsize": size * 0.75,
    "figure.titlesize": size * 1.5,
    "axes.titlesize": size * 1.5,
    "axes.titlepad": size,
    "axes.labelpad": size - 10,
    "lines.linewidth": 2,
    "axes.spines.top": False,
    "axes.spines.right": False,
    "axes.spines.left": False,
    "axes.spines.bottom": False,
    "legend.fontsize": size,
    "figure.figsize": (10, 6),
}

# **Imputing Missing Values**

functions.MissingValues(data)

## **Home Planet**

### **Impute Home Planet from Group**

unique_homeplanets_per_group = data.groupby('Group')['HomePlanet'].nunique()
groups_with_multiple_planets = unique_homeplanets_per_group[unique_homeplanets_per_group > 1]

print(groups_with_multiple_planets)

group_homeplanet_map = data.groupby('Group')['HomePlanet'].apply(lambda x: x.mode()[0] if not x.mode().empty else None)

def fill_missing_homeplanet(row):
    if pd.isna(row['HomePlanet']):
        return group_homeplanet_map.get(row['Group'], row['HomePlanet'])
    else:
        return row['HomePlanet']

data['HomePlanet'] = data.apply(fill_missing_homeplanet, axis=1)

functions.MissingValues(data)

### **Impute Home Planet from Last Name**

lastname_homeplanet_map = data.groupby('LastName')['HomePlanet'].apply(lambda x: x.mode()[0] if not x.mode().empty else None)

def fill_missing_homeplanet_by_lastname(row):
    if pd.isna(row['HomePlanet']):
        return lastname_homeplanet_map.get(row['LastName'], row['HomePlanet'])
    else:
        return row['HomePlanet']

data['HomePlanet'] = data.apply(fill_missing_homeplanet_by_lastname, axis=1)

functions.MissingValues(data)

### **Impute Home Planet from Mode**

homeplanet_mode = data['HomePlanet'].mode()[0]
print(homeplanet_mode)

mode_value = data['HomePlanet'].mode()

if not mode_value.empty:
    data['HomePlanet'] = data['HomePlanet'].fillna(mode_value[0])

functions.MissingValues(data)

### **Insights**

>  **0. The Code:** Grouping HomePlanet by Group/LastName. Retrieving the most common HomePlanet (mode) based on Group/LastName</BR>
>  **1. Grouping by Group:** Starting with 288 missing values. Grouping by 'Group' to find the most common (or only) HomePlanet. Fill missing HomePlanet values based on Group's identified home planet. End with 157 missing values. </BR>
>  **2. Grouping by Last Name:** Starting with 157 missing values. Grouping by 'LastName' to find the most common (or only) HomePlanet. Fill missing HomePlanet values based on last name's identified home planet. End with 7 missing values. </BR>
>  **3. Mode:** Start with 7 missing values. Fill missing values by mode of HomePlanet, which is Earth. End with 0 missing values.

## **Destination**

### **Impute Destination from Group**

unique_destinationplanets_per_group = data.groupby('Group')['Destination'].nunique()
groups_with_multiple_destinations = unique_destinationplanets_per_group[unique_destinationplanets_per_group > 1]

groups_with_multiple_destinations.sort_values(ascending=True)

group_destinationplanet_map = data.groupby('Group')['Destination'].apply(lambda x: x.mode()[0] if not x.mode().empty else None)

def fill_missing_destinationplanet(row):
    if pd.isna(row['Destination']):
        return group_destinationplanet_map.get(row['Group'], row['Destination'])
    else:
        return row['Destination']

data['Destination'] = data.apply(fill_missing_destinationplanet, axis=1)

functions.MissingValues(data)

### **Impute Destination from Last Name**

lastname_destinationplanet_map = data.groupby('LastName')['Destination'].apply(lambda x: x.mode()[0] if not x.mode().empty else None)

def fill_missing_destinationplanet_by_lastname(row):
    if pd.isna(row['Destination']):
        return lastname_destinationplanet_map.get(row['LastName'], row['Destination'])
    else:
        return row['Destination']

data['Destination'] = data.apply(fill_missing_destinationplanet_by_lastname, axis=1)

functions.MissingValues(data)

### **Impute Destination from Mode**

desinationplanet_mode = data['Destination'].mode()[0]
print(desinationplanet_mode)

data['Destination'] = data['Destination'].fillna(data['Destination'].mode()[0])

functions.MissingValues(data)

>  **0. The Code:** Grouping Destination by Group/LastName. Retrieving the most common Destination (mode) based on Group/LastName</BR>
>  **1. Grouping by Group:** Starting with 274 missing values. Grouping by 'Group' to find the most common (or only) destination. Fill missing values based on Group's identified destination. End with 154 missing values. </BR>
>  **2. Grouping by Last Name:** Starting with 154 missing values. Grouping by 'LastName' to find the most common (or only) destination. Fill missing values based on last name's identified destination. End with 4 missing values. </BR>
>  **3. Mode:** Start with 4 missing values. Fill missing values by mode, which is TRAPPIST-1e. End with 0 missing values.

## **Cabin Deck**

### **Cabin Deck and Home Planet**

data.pivot_table(index='CabinDeck', columns='HomePlanet', aggfunc='size', fill_value=0)

### **Impute CabinDeck G from Earth**

data.loc[(data['HomePlanet'] == 'Earth') & (data['CabinDeck'].isnull()), 'CabinDeck'] = 'G'

functions.MissingValues(data)

### **Impute CabinDeck from Last Name**

lastname_cabindeck_map = data.groupby('LastName')['CabinDeck'].apply(lambda x: x.mode()[0] if not x.mode().empty else None)

def fill_missing_cabindeck_by_lastname(row):
    if pd.isna(row['CabinDeck']):
        return lastname_homeplanet_map.get(row['LastName'], row['CabinDeck'])
    else:
        return row['CabinDeck']

data['CabinDeck'] = data.apply(fill_missing_cabindeck_by_lastname, axis=1)

functions.MissingValues(data)

>  **1. Decks and Home Planets:** Starting with 299 missing values. Passengers on decks A (Except 1 passenger), B, C (Except 1 passenger), T came from Europa. Passengers on deck G came from Earth. Passengers on decks D, E or F came from multiple planets. </BR>
>  **2. Earth and Cabin Deck G:** Fill missing CabinDeck 'G' where HomePlanet is 'Earth'. End with 138 missing values. </BR> 
>  **3. Grouping by Last Name:** Starting with 138 missing values. Grouping by 'LastName' to find the most common (or only) cabin deck. Fill missing values based on last name's identified cabin deck. End with 0 missing values. </BR>

## **Cabin Side**

unique_cabinside_per_group = data[data['GroupSize']>1].groupby('Group')['CabinSide'].nunique()
grouped_by_unique_count = unique_cabinside_per_group.value_counts().sort_index()

grouped_by_unique_count = grouped_by_unique_count.to_frame().reset_index()

### **Cabin Side from Group**

group_cabinside = data[data['GroupSize'] > 1].groupby(['Group', 'CabinSide']).size().unstack(fill_value=0)

missing_cabinside_mask = data['CabinSide'].isna() & data['Group'].isin(group_cabinside.index)

data.loc[missing_cabinside_mask, 'CabinSide'] = data.loc[missing_cabinside_mask, 'Group'].map(group_cabinside.idxmax(axis=1))

functions.MissingValues(data)

### **Cabin Side from Last Name**

group_cabinside = data[data['GroupSize'] > 1].groupby(['LastName', 'CabinSide']).size().unstack(fill_value=0)

missing_cabinside_mask = data['CabinSide'].isna() & data['LastName'].isin(group_cabinside.index)

data.loc[missing_cabinside_mask, 'CabinSide'] = data.loc[missing_cabinside_mask, 'LastName'].map(group_cabinside.idxmax(axis=1))

functions.MissingValues(data)

### **Replace Cabin Side Missing Values by Letter U**

# Missing values before
CS_bef=data['CabinSide'].isna().sum()

# Fill remaining missing values with outlier
data.loc[data['CabinSide'].isna(),'CabinSide']='U'

functions.MissingValues(data)

### **Insights**

>  **1. Grouping by Group:** Starting with 299 missing values. Everyone in the same group is also on the same cabin side. Grouping by 'Group' to find the most common (or only) cabin side. Fill missing values based on Group's identified cabin side. End with 162 missing values. </BR>
>  **2. Grouping by Last Name:** Starting with 162 missing values. Grouping by 'LastName' to find the most common (or only) cabin side. Fill missing values based on last name's identified destination. End with 65 missing values. </BR>
>  **3. Fill by U (Unknown):** Starting with 65 missing values. They will be replaced with letter U (Unknown). This is because we really don't know which one of the two (balanced) sides we should assign. End with 0 missing values.

## **Cabin Number**

functions.MissingValues(data)

data['CabinNumber'] = data.groupby(['Group'])['CabinNumber'].transform(lambda x: x.fillna(x.mode()[0] if not x.mode().empty else 'Unknown'))

functions.MissingValues(data)

### **Insights**

>  **1. Starting Missing Values:** Starting with 299 missing values.</BR>
>  **2. Linear Regrerssion:** Extrapolate the missing cabin numbers using mode. End with 0 missing values.

## **CryoSleep**

pd.set_option('future.no_silent_downcasting', True)

data['CryoSleep'] = data['CryoSleep'].fillna(False).astype(bool)

functions.MissingValues(data)

### **Insights**

>  **1. Starting Missing Values:** Starting with 310 missing values.</BR>
>  **2. CryoSleep == False:** Imputed Missing values with False since most passengers were False. </BR>
>  **3. End Missing Values:** End with 0 missing values.

## **VIP**

data['VIP'] = data['VIP'].fillna(False).astype(bool)

functions.MissingValues(data)

### **Insights**

>  **1. Starting Missing Values:** Starting with 3296 missing values.</BR>
>  **2. VIP == False:** Imputed Missing values with False since most passengers were False. </BR>
>  **3. End Missing Values:** End with 0 missing values.

## **Save Dataset**

data.to_csv(r"C:\Users\Dell\Documents\AI\Titanic\Data\Data\data.csv", index=False)
