==> Project/06 AutoML.ipynb <==
# <center> **Kaggle’s Spaceship Titanic Competition**
# <center> **Machine Learning**

# **Libraries**

import pandas as pd
import numpy as np

from sklearn.preprocessing import StandardScaler
from sklearn.preprocessing import OneHotEncoder
from sklearn.utils import shuffle
from sklearn.model_selection import train_test_split

import functions
import importlib
importlib.reload(functions)

import lazypredict
from lazypredict.Supervised import LazyClassifier

import time

# **Load Data**

data = pd.read_csv(
    r"C:\Users\Dell\Documents\AI\Titanic\Data\Data\data.csv",
    index_col=False
)

train = pd.read_csv(
    r"C:\Users\Dell\Documents\AI\Titanic\Data\train.csv",
    index_col=False
)

test = pd.read_csv(
    r"C:\Users\Dell\Documents\AI\Titanic\Data\test.csv",
    index_col=False
)


random_state = 101
target = 'Transported'

# **Machine Learning**

## **Split Data Back to Train and Test**

train=data[data['PassengerId'].isin(train['PassengerId'].values)].copy()
test=data[data['PassengerId'].isin(test['PassengerId'].values)].copy()

## **Drop Unneeded Features**

train.drop(['Group', 'CabinNumber'], axis=1, inplace=True)
test.drop(['Group', 'CabinNumber'], axis=1, inplace=True)

## **Pre-Processing**

X = train.drop(target, axis=1)
y = train[target]
y = y.astype(bool)

numerical_cols = [cname for cname in X.columns if X[cname].dtype in ['int64', 'float64']]
categorical_cols = [cname for cname in X.columns if X[cname].dtype in ["object", "bool"]]

scaler = StandardScaler()
scaled_numerical = scaler.fit_transform(X[numerical_cols])

encoder = OneHotEncoder(handle_unknown='ignore', sparse_output=False)  # Set sparse_output=False to get dense output
encoded_categorical = encoder.fit_transform(X[categorical_cols])

scaled_numerical_df = pd.DataFrame(scaled_numerical, columns=numerical_cols)

encoded_categorical_df = pd.DataFrame(encoded_categorical, columns=encoder.get_feature_names_out(categorical_cols))

scaled_numerical_df.reset_index(drop=True, inplace=True)
encoded_categorical_df.reset_index(drop=True, inplace=True)

X = pd.concat([scaled_numerical_df, encoded_categorical_df], axis=1)

## **Train Test Split**

X, y = shuffle(X, y, random_state=random_state)
X_train, X_test, y_train, y_test = train_test_split(X, y, test_size=0.2, stratify=y, random_state=random_state)

## **LazyPredict**

clf = LazyClassifier(verbose=0,ignore_warnings=True)
models, predictions = clf.fit(X_train, X_test, y_train, y_test)
models

### **Insights**

> **1. LazyPredict Time to Execute:** 4 Hours. </BR>
> **2. Top 3 Modles:** LGBM Classifier, XGB Classifier, Bagging Classifier </BR>
> **3. Best Accuracy Score:** 0.80 </BR>
> **4. Conclusion:** Results are comparable to results from previous notebook </BR>
