==> Project/07 Conclusions.ipynb <==
# <center> **Kaggle’s Spaceship Titanic Competition**
# <center> **Conclusions**
# <center> **Suggestions for Improvement**

# **Conclusions**

Here is a summary of the conclusions that may be drawn from this report. 

>* **The Analysis of the Data:** I reviewed approximately 8,700 datapoint related to passnegers on the Spaceship Titanic. <br> 
>* **The Goal of the Project:** The goal of this project was to find a model that could predict if a passenger was transported to an alternative dimensions with an accuracy score over 79.<br>
>* **Missing Data:** There was over 2% missing data in every feature except PassengerId. However, I was able to impute some of this missing information using known data from other features. For example, from Last Name of the passenger it was possible to impute some of the Home Planet and some of the Destination that was missing.  <br>
>* **Croygenic Sleep:** Passengers who were in Cryogenic Sleep during the trip were less likely to be transported. Cryogenic Sleep turend out to be a very important feature. <br>
>* **Luxury Spending:** Passengers who spend money on luxury items like spa were more likely to be transported. Expenditure on luxury items turned out important in predicting Transportation.<br>
>* **Models:** I utilized the following models: Logistic Regression, K Nearest Neighbors, Random Forest, Extreme Gradient Boosting (XGB), Light Gradient Boosting Machine (LGBM), Categorical Boosting (CatBoost). The boosting models gave the best performance. I used GridSearch CV with the three boosting models to tune their hyperparameters. <br>
>* **Auto ML:** I utilized LazyPredict. Top 3 models were LGBM Classifier, XGB Classifier and Bagging Classifier. These results are comparable to the results that I obtained by manual machine learning, but the process was extremely slow at 4 hours. <br>
>* **Extremet Gradient Boosting (XGB):** Best Cross Validation Score: 0.81. Best Test Score: 0.78. Modeling Time: 22 minutes   <br>
>* **Light Gradient Boosting (LGBM):** Best Cross Validation Score: 0.81. Best Test Score: 0.79. Modeling Time: 2 minutes.  <br>
>* **Categorical Boosting (CatBoost):** Best Cross Validation Score: 0.81. Best Test Score: 0.78. Modeling Time: 71 minutes.<br>
>* **Final Model (LGBM):** I used LGBM as my final model. I obtained a score of 0.80032 on the Kaggle competition.<br>

# **Suggestions for Improvement**

This report has certain weaknesses. In this section, I outlined those weaknesses and indicated some avenues for improvement. 

>* **Domain Knowledge:** It is best if the data scientist has adequate domain knowledge on the topic of the analysis. I do not have any expertise in space travel or alternative dimensions. There may be parts of the data that I have overlooked that may have been important and I may have given importance to parts that may have had little significance. <br>
>* **More Complete Data:** As mentioned earlier, there was missing data in every feature. Less missing data could have improved the performance of the models.<br>  
>* **Pandas:** Continue to learn to utilize more optimized Pandas techniques and algorithms.<br>
>* **Seaborn and Matplotlib:** Continue to improve my knowledge of Seaborn and Matplotlib for creating visualizations. <br>
>* **Machine Learning:** Continue to improve on my capabilities with various models. <br>
>* **Python Code:** Continue to write better and more efficient Python code. <br>
>* **Clean Code:** Continue to adhere to the principles of writing clean code. <br>
>* **Readability and Efficiency:** Continue to improve my skills to find the delicate balance between readability and efficiency in coding.<br>
